fix bug in percentage

+++ libhdt/src/util/Domains.hpp
@@ -411,7 +411,7 @@ public:
 	}
 
 	double getWeightShared(){
-		double ret = (totalSubjectObjects+totalOtherLessThreshold_shared)/((totalUniqueSubjects+totalOtherLessThreshold_UniqueSubjects)+(totalSubjectObjects+totalOtherLessThreshold_shared));
+		double ret = (double(totalSubjectObjects+totalOtherLessThreshold_shared))/((totalUniqueSubjects+totalOtherLessThreshold_UniqueSubjects)+(totalSubjectObjects+totalOtherLessThreshold_shared));
 		return ret;
 	}
